Move Retrofit to api package
