+++ src/main/java/io/jenkins/plugins/appcenter/task/AppCenterTask.java
@@ -1,9 +1,9 @@
 package io.jenkins.plugins.appcenter.task;
 
 import io.jenkins.plugins.appcenter.AppCenterException;
-import io.jenkins.plugins.appcenter.model.AppCenterService;
-import io.jenkins.plugins.appcenter.model.AppCenterServiceFactory;
-import io.jenkins.plugins.appcenter.model.UploadService;
+import io.jenkins.plugins.appcenter.api.AppCenterService;
+import io.jenkins.plugins.appcenter.api.AppCenterServiceFactory;
+import io.jenkins.plugins.appcenter.api.UploadService;
 import jenkins.security.MasterToSlaveCallable;
 
 import java.io.IOException;
